Form Creation

+++ orderForm/src/com/xworkz/dto/CustomerDTO.java
@@ -6,18 +6,18 @@ import java.time.LocalDate;
 public class CustomerDTO implements Serializable {
     private String name;
     private String email;
-    private double mobile;
+    private long mobile;
     private String house;
     private String area;
     private String city;
-    private double pincode;
+    private long pincode;
     private LocalDate date;
 
     public CustomerDTO(){
         System.out.println("No args Customer DTO Constructor........");
     }
 
-    public CustomerDTO(String name, String email, double mobile, String house, String area, String city, double pincode, LocalDate date) {
+    public CustomerDTO(String name, String email, long mobile, String house, String area, String city, long pincode, LocalDate date) {
         this.name = name;
         this.email = email;
         this.mobile = mobile;
@@ -44,7 +44,7 @@ public class CustomerDTO implements Serializable {
         return mobile;
     }
 
-    public void setMobile(double mobile) {
+    public void setMobile(long mobile) {
         this.mobile = mobile;
     }
 
@@ -76,7 +76,7 @@ public class CustomerDTO implements Serializable {
         return pincode;
     }
 
-    public void setPincode(double pincode) {
+    public void setPincode(long pincode) {
         this.pincode = pincode;
     }
